backend/requirements.txt
fastapi>=0.115.0
uvicorn>=0.30.0
google-api-python-client>=2.0.0
python-dotenv>=1.0.0
llama-cpp-python>=0.3.0
huggingface_hub>=0.24.0
